@@ -12,16 +12,16 @@ def custom_format(compilation_unit: SlitherCompilationUnit, result):
                 element["type_specific_fields"]["parent"]["name"]
             )
             if target_contract:
-                function = target_contract.get_function_from_signature(
+                function = target_contract.get_function_from_full_name(
                     element["type_specific_fields"]["signature"]
                 )
                 if function:
                     _patch(
-                        compilation_unit,
+                        file_scope,
                         result,
                         element["source_mapping"]["filename_absolute"],
-                        int(function.parameters_src().source_mapping["start"]),
-                        int(function.returns_src().source_mapping["start"]),
+                        int(function.parameters_src().source_mapping.start),
+                        int(function.returns_src().source_mapping.start),
                     )
 
 
