@@ -1,6 +1,8 @@
 import re
 import logging
 
+from typing import List
+
 from slither.core.compilation_unit import SlitherCompilationUnit
 from slither.slithir.operations import (
     Send,
@@ -10,6 +12,7 @@ from slither.slithir.operations import (
     LowLevelCall,
     InternalCall,
     InternalDynamicCall,
+    Operation,
 )
 from slither.core.declarations import Modifier
 from slither.core.solidity_types import UserDefinedType, MappingType
@@ -254,7 +257,7 @@ def _patch(compilation_unit: SlitherCompilationUnit, result, element, _target):
         ]
         param_name = element["name"]
         contract = scope.get_contract_from_name(contract_name)
-        function = contract.get_function_from_signature(function_sig)
+        function = contract.get_function_from_full_name(function_sig)
         target = function.get_local_variable_from_name(param_name)
 
     elif _target in ["variable", "variable_constant"]:
@@ -268,7 +271,7 @@ def _patch(compilation_unit: SlitherCompilationUnit, result, element, _target):
             ]
             var_name = element["name"]
             contract = scope.get_contract_from_name(contract_name)
-            function = contract.get_function_from_signature(function_sig)
+            function = contract.get_function_from_full_name(function_sig)
             target = function.get_local_variable_from_name(var_name)
         # State variable
         else:
@@ -300,10 +303,10 @@ def _patch(compilation_unit: SlitherCompilationUnit, result, element, _target):
 # group 2: beginning of the to type
 # nested mapping are within the group 1
 # RE_MAPPING = '[ ]*mapping[ ]*\([ ]*([\=\>\(\) a-zA-Z0-9\._\[\]]*)[ ]*=>[ ]*([a-zA-Z0-9\._\[\]]*)\)'
-RE_MAPPING_FROM = b"([a-zA-Z0-9\._\[\]]*)"
-RE_MAPPING_TO = b"([\=\>\(\) a-zA-Z0-9\._\[\]\   ]*)"
+RE_MAPPING_FROM = rb"([a-zA-Z0-9\._\[\]]*)"
+RE_MAPPING_TO = rb"([\=\>\(\) a-zA-Z0-9\._\[\]\   ]*)"
 RE_MAPPING = (
-    b"[ ]*mapping[ ]*\([ ]*" + RE_MAPPING_FROM + b"[ ]*" + b"=>" + b"[ ]*" + RE_MAPPING_TO + b"\)"
+    rb"[ ]*mapping[ ]*\([ ]*" + RE_MAPPING_FROM + b"[ ]*" + b"=>" + b"[ ]*" + RE_MAPPING_TO + rb"\)"
 )
 
 
@@ -418,9 +421,9 @@ def _explore_variables_declaration(  # pylint: disable=too-many-arguments,too-ma
 ):
     for variable in variables:
         # First explore the type of the variable
-        filename_source_code = variable.source_mapping["filename_absolute"]
-        full_txt_start = variable.source_mapping["start"]
-        full_txt_end = full_txt_start + variable.source_mapping["length"]
+        filename_source_code = variable.source_mapping.filename.absolute
+        full_txt_start = variable.source_mapping.start
+        full_txt_end = full_txt_start + variable.source_mapping.length
         full_txt = slither.source_code[filename_source_code].encode("utf8")[
             full_txt_start:full_txt_end
         ]
@@ -433,7 +436,7 @@ def _explore_variables_declaration(  # pylint: disable=too-many-arguments,too-ma
             variable.type,
             filename_source_code,
             full_txt_start,
-            variable.source_mapping["start"] + variable.source_mapping["length"],
+            variable.source_mapping.start + variable.source_mapping.length,
         )
 
         # If the variable is the target
@@ -448,9 +451,9 @@ def _explore_variables_declaration(  # pylint: disable=too-many-arguments,too-ma
 
             # Patch comment only makes sense for local variable declaration in the parameter list
             if patch_comment and isinstance(variable, LocalVariable):
-                if "lines" in variable.source_mapping and variable.source_mapping["lines"]:
+                if variable.source_mapping.lines:
                     func = variable.function
-                    end_line = func.source_mapping["lines"][0]
+                    end_line = func.source_mapping.lines[0]
                     if variable in func.parameters:
                         idx = len(func.parameters) - func.parameters.index(variable) + 1
                         first_line = end_line - idx - 2
@@ -462,8 +465,8 @@ def _explore_variables_declaration(  # pylint: disable=too-many-arguments,too-ma
                             first_line : end_line - 1
                         ]
 
-                        idx_beginning = func.source_mapping["start"]
-                        idx_beginning += -func.source_mapping["starting_column"] + 1
+                        idx_beginning = func.source_mapping.start
+                        idx_beginning += -func.source_mapping.starting_column + 1
                         idx_beginning += -sum([len(c) for c in potential_comments])
 
                         old_comment = f"@param {old_str}".encode("utf8")
@@ -498,9 +501,9 @@ def _explore_structures_declaration(slither, structures, result, target, convert
             old_str = st.name
             new_str = convert(old_str, slither)
 
-            filename_source_code = st.source_mapping["filename_absolute"]
-            full_txt_start = st.source_mapping["start"]
-            full_txt_end = full_txt_start + st.source_mapping["length"]
+            filename_source_code = st.source_mapping.filename.absolute
+            full_txt_start = st.source_mapping.start
+            full_txt_end = full_txt_start + st.source_mapping.length
             full_txt = slither.source_code[filename_source_code].encode("utf8")[
                 full_txt_start:full_txt_end
             ]
@@ -521,12 +524,12 @@ def _explore_events_declaration(slither, events, result, target, convert):
 
         # If the event is the target
         if event == target:
-            filename_source_code = event.source_mapping["filename_absolute"]
+            filename_source_code = event.source_mapping.filename.absolute
 
             old_str = event.name
             new_str = convert(old_str, slither)
 
-            loc_start = event.source_mapping["start"]
+            loc_start = event.source_mapping.start
             loc_end = loc_start + len(old_str)
 
             create_patch(result, filename_source_code, loc_start, loc_end, old_str, new_str)
@@ -550,7 +553,7 @@ def get_ir_variables(ir):
     return [v for v in all_vars if v]
 
 
-def _explore_irs(slither, irs, result, target, convert):
+def _explore_irs(slither, irs: List[Operation], result, target, convert):
     # pylint: disable=too-many-locals
     if irs is None:
         return
@@ -562,9 +565,9 @@ def _explore_irs(slither, irs, result, target, convert):
                 and v.canonical_name == target.canonical_name
             ):
                 source_mapping = ir.expression.source_mapping
-                filename_source_code = source_mapping["filename_absolute"]
-                full_txt_start = source_mapping["start"]
-                full_txt_end = full_txt_start + source_mapping["length"]
+                filename_source_code = source_mapping.filename.absolute
+                full_txt_start = source_mapping.start
+                full_txt_end = full_txt_start + source_mapping.length
                 full_txt = slither.source_code[filename_source_code].encode("utf8")[
                     full_txt_start:full_txt_end
                 ]
@@ -606,9 +609,9 @@ def _explore_functions(slither, functions, result, target, convert):
             old_str = function.name
             new_str = convert(old_str, slither)
 
-            filename_source_code = function.source_mapping["filename_absolute"]
-            full_txt_start = function.source_mapping["start"]
-            full_txt_end = full_txt_start + function.source_mapping["length"]
+            filename_source_code = function.source_mapping.filename.absolute
+            full_txt_start = function.source_mapping.start
+            full_txt_end = full_txt_start + function.source_mapping.length
             full_txt = slither.source_code[filename_source_code].encode("utf8")[
                 full_txt_start:full_txt_end
             ]
@@ -631,9 +634,9 @@ def _explore_enums(slither, enums, result, target, convert):
             old_str = enum.name
             new_str = convert(old_str, slither)
 
-            filename_source_code = enum.source_mapping["filename_absolute"]
-            full_txt_start = enum.source_mapping["start"]
-            full_txt_end = full_txt_start + enum.source_mapping["length"]
+            filename_source_code = enum.source_mapping.filename.absolute
+            full_txt_start = enum.source_mapping.start
+            full_txt_end = full_txt_start + enum.source_mapping.length
             full_txt = slither.source_code[filename_source_code].encode("utf8")[
                 full_txt_start:full_txt_end
             ]
@@ -654,9 +657,9 @@ def _explore_contract(slither, contract, result, target, convert):
     _explore_enums(slither, contract.enums, result, target, convert)
 
     if contract == target:
-        filename_source_code = contract.source_mapping["filename_absolute"]
-        full_txt_start = contract.source_mapping["start"]
-        full_txt_end = full_txt_start + contract.source_mapping["length"]
+        filename_source_code = contract.source_mapping.filename.absolute
+        full_txt_start = contract.source_mapping.start
+        full_txt_end = full_txt_start + contract.source_mapping.length
         full_txt = slither.source_code[filename_source_code].encode("utf8")[
             full_txt_start:full_txt_end
         ]
