@@ -11,6 +11,7 @@ from slither.core.declarations.function import (
 )
 from slither.core.declarations.function_contract import FunctionContract
 from slither.core.expressions import AssignmentOperation
+from slither.core.source_mapping.source_mapping import Source
 from slither.core.variables.local_variable import LocalVariable
 from slither.core.variables.local_variable_init_from_tuple import LocalVariableInitFromTuple
 from slither.solc_parsing.cfg.node import NodeSolc
@@ -319,7 +320,7 @@ class FunctionSolc(CallerContextExpression):
     ###################################################################################
 
     def _new_node(
-        self, node_type: NodeType, src: Union[str, Dict], scope: Union[Scope, "Function"]
+        self, node_type: NodeType, src: Union[str, Source], scope: Union[Scope, "Function"]
     ) -> NodeSolc:
         node = self._function.new_node(node_type, src, scope)
         node_parser = NodeSolc(node)
@@ -1347,7 +1348,7 @@ class FunctionSolc(CallerContextExpression):
                     condition = st.condition
                     if not condition:
                         raise ParsingError(
-                            f"Incorrect ternary conversion {node.expression} {node.source_mapping_str}"
+                            f"Incorrect ternary conversion {node.expression} {node.source_mapping}"
                         )
                     true_expr = st.true_expression
                     false_expr = st.false_expression
