@@ -15,13 +15,13 @@ def remove_assignement(variable: Variable, contract: Contract, result: Dict):
     :return:
     """
     # Retrieve the file
-    in_file = contract.source_mapping["filename_absolute"]
+    in_file = contract.source_mapping.filename.absolute
     # Retrieve the source code
     in_file_str = contract.compilation_unit.core.source_code[in_file]
 
     # Get the string
-    start = variable.source_mapping["start"]
-    stop = variable.expression.source_mapping["start"]
+    start = variable.source_mapping.start
+    stop = variable.expression.source_mapping.start
     old_str = in_file_str[start:stop]
 
     new_str = old_str[: old_str.find("=")]
@@ -30,7 +30,7 @@ def remove_assignement(variable: Variable, contract: Contract, result: Dict):
         result,
         in_file,
         start,
-        stop + variable.expression.source_mapping["length"],
+        stop + variable.expression.source_mapping.length,
         old_str,
         new_str,
     )
