@@ -160,7 +160,7 @@ def output_to_sarif(
         ],
     }
 
-    for detector in results["detectors"]:
+    for detector in results.get("detectors", []):
         _output_result_to_sarif(detector, detectors_classes, sarif)
 
     if filename == "-":
@@ -228,14 +228,14 @@ def _convert_to_description(d):
 
     if isinstance(d, Node):
         if d.expression:
-            return f"{d.expression} ({d.source_mapping_str})"
-        return f"{str(d)} ({d.source_mapping_str})"
+            return f"{d.expression} ({d.source_mapping})"
+        return f"{str(d)} ({d.source_mapping})"
 
     if hasattr(d, "canonical_name"):
-        return f"{d.canonical_name} ({d.source_mapping_str})"
+        return f"{d.canonical_name} ({d.source_mapping})"
 
     if hasattr(d, "name"):
-        return f"{d.name} ({d.source_mapping_str})"
+        return f"{d.name} ({d.source_mapping})"
 
     raise SlitherError(f"{type(d)} cannot be converted (no name, or canonical_name")
 
@@ -249,14 +249,14 @@ def _convert_to_markdown(d, markdown_root):
 
     if isinstance(d, Node):
         if d.expression:
-            return f"[{d.expression}]({d.source_mapping_to_markdown(markdown_root)})"
-        return f"[{str(d)}]({d.source_mapping_to_markdown(markdown_root)})"
+            return f"[{d.expression}]({d.source_mapping.to_markdown(markdown_root)})"
+        return f"[{str(d)}]({d.source_mapping.to_markdown(markdown_root)})"
 
     if hasattr(d, "canonical_name"):
-        return f"[{d.canonical_name}]({d.source_mapping_to_markdown(markdown_root)})"
+        return f"[{d.canonical_name}]({d.source_mapping.to_markdown(markdown_root)})"
 
     if hasattr(d, "name"):
-        return f"[{d.name}]({d.source_mapping_to_markdown(markdown_root)})"
+        return f"[{d.name}]({d.source_mapping.to_markdown(markdown_root)})"
 
     raise SlitherError(f"{type(d)} cannot be converted (no name, or canonical_name")
 
@@ -275,11 +275,11 @@ def _convert_to_id(d):
 
     if isinstance(d, Node):
         if d.expression:
-            return f"{d.expression} ({d.source_mapping_str})"
-        return f"{str(d)} ({d.source_mapping_str})"
+            return f"{d.expression} ({d.source_mapping})"
+        return f"{str(d)} ({d.source_mapping})"
 
     if isinstance(d, Pragma):
-        return f"{d} ({d.source_mapping_str})"
+        return f"{d} ({d.source_mapping})"
 
     if hasattr(d, "canonical_name"):
         return f"{d.canonical_name}"
@@ -299,7 +299,7 @@ def _convert_to_id(d):
 
 
 def _create_base_element(
-    custom_type, name, source_mapping, type_specific_fields=None, additional_fields=None
+    custom_type, name, source_mapping: Dict, type_specific_fields=None, additional_fields=None
 ):
     if additional_fields is None:
         additional_fields = {}
@@ -380,7 +380,7 @@ class Output:
 
     def add(self, add: SupportedOutput, additional_fields: Optional[Dict] = None):
         if not self._data["first_markdown_element"]:
-            self._data["first_markdown_element"] = add.source_mapping_to_markdown(
+            self._data["first_markdown_element"] = add.source_mapping.to_markdown(
                 self._markdown_root
             )
         if isinstance(add, Variable):
@@ -424,7 +424,7 @@ class Output:
         element = _create_base_element(
             "variable",
             variable.name,
-            variable.source_mapping,
+            variable.source_mapping.to_json(),
             type_specific_fields,
             additional_fields,
         )
@@ -445,7 +445,7 @@ class Output:
         if additional_fields is None:
             additional_fields = {}
         element = _create_base_element(
-            "contract", contract.name, contract.source_mapping, {}, additional_fields
+            "contract", contract.name, contract.source_mapping.to_json(), {}, additional_fields
         )
         self._data["elements"].append(element)
 
@@ -466,7 +466,7 @@ class Output:
         element = _create_base_element(
             "function",
             function.name,
-            function.source_mapping,
+            function.source_mapping.to_json(),
             type_specific_fields,
             additional_fields,
         )
@@ -492,7 +492,7 @@ class Output:
         element = _create_base_element(
             "enum",
             enum.name,
-            enum.source_mapping,
+            enum.source_mapping.to_json(),
             type_specific_fields,
             additional_fields,
         )
@@ -512,7 +512,7 @@ class Output:
         element = _create_base_element(
             "struct",
             struct.name,
-            struct.source_mapping,
+            struct.source_mapping.to_json(),
             type_specific_fields,
             additional_fields,
         )
@@ -535,7 +535,7 @@ class Output:
         element = _create_base_element(
             "event",
             event.name,
-            event.source_mapping,
+            event.source_mapping.to_json(),
             type_specific_fields,
             additional_fields,
         )
@@ -559,7 +559,7 @@ class Output:
         element = _create_base_element(
             "node",
             node_name,
-            node.source_mapping,
+            node.source_mapping.to_json(),
             type_specific_fields,
             additional_fields,
         )
@@ -583,7 +583,7 @@ class Output:
         element = _create_base_element(
             "pragma",
             pragma.version,
-            pragma.source_mapping,
+            pragma.source_mapping.to_json(),
             type_specific_fields,
             additional_fields,
         )
@@ -667,7 +667,7 @@ class Output:
 
         # If this is a source mapping object, get the underlying source mapping dictionary
         if isinstance(source_mapping, SourceMapping):
-            source_mapping = source_mapping.source_mapping
+            source_mapping = source_mapping.source_mapping.to_json()
 
         # Create the underlying element and add it to our resulting json
         element = _create_base_element("other", name, source_mapping, {}, additional_fields)
